[#42] refactor: 리뷰 관련 로직 분리
