+++ src/main/java/com/itda/apiserver/controller/ProductController.java
@@ -2,7 +2,6 @@ package com.itda.apiserver.controller;
 
 import com.itda.apiserver.annotation.LoginRequired;
 import com.itda.apiserver.annotation.UserId;
-import com.itda.apiserver.dto.AddReviewRequestDto;
 import com.itda.apiserver.dto.AddproductRequestDto;
 import com.itda.apiserver.dto.ApiResult;
 import com.itda.apiserver.dto.GetAllProductDto;
@@ -29,14 +28,6 @@ public class ProductController {
     @LoginRequired
     public ApiResult<Void> addProduct(@RequestBody AddproductRequestDto addproductRequestDto, @UserId Long userId) {
         productService.addProduct(addproductRequestDto, userId);
-        return ApiResult.ok(null);
-    }
-
-    @PostMapping("/{productId}/review")
-    @LoginRequired
-    public ApiResult<Void> addReview(@PathVariable Long productId, @UserId Long userId,
-                                     @RequestBody AddReviewRequestDto addReviewRequest) {
-        productService.addReview(userId, productId, addReviewRequest);
         return ApiResult.ok(null);
     }
 }
